extract components from backups page
