+++ src/app/instances/[name]/backups/types/backup.ts
@@ -1,0 +1,7 @@
+export default interface Backup {
+  timestamp: string;
+  rabbitmq_version: string;
+  trigger: string;
+  definitions: Record<string, unknown>;
+}
+
